"Adds navigation and return button to LeaderBoard component, adjusts responsive layout in QuizList component
This commit adds the ability to navigate back to the dashboard from the LeaderBoard component by including a "Return To Dashboard" button. In addition, the responsive layout in the QuizList component has been adjusted to display 2 columns on screens with a maximum width of 992px. These changes improve the user experience and make it easier to navigate between different sections of the application."